server/requirements.txt
click==8.0.4
colorama==0.4.4
Cython==0.29.28
Flask==2.0.3
Flask-JWT-Extended==4.3.1
hurst==0.0.5
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
patsy==0.5.2
pmdarima==1.8.5
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
simpy==4.0.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.0.3
